refactor: repoActiInflTrend
