+++ src/views/RepoActiInflTrend/index.tsx
@@ -1,16 +1,7 @@
 import React, { useState, useEffect } from 'react';
-import { render } from 'react-dom';
-import $ from 'jquery';
-import * as pageDetect from 'github-url-detection';
-import {
-  getGithubTheme,
-  runsWhen,
-  getMessageByLocale,
-} from '../../utils/utils';
-import PerceptorBase from './PerceptorBase';
-import { inject2Perceptor } from './Perceptor';
+
+import { getGithubTheme, getMessageByLocale } from '../../utils/utils';
 import Settings, { loadSettings } from '../../utils/settings';
-import { utils } from 'github-url-detection';
 import { getRepoActiInfl } from '../../api/repo';
 import Bars from '../../components/Bars/index';
 
@@ -112,48 +103,4 @@ const RepoActiInflTrendView: React.FC<RepoActiInflTrendViewProps> = ({
   );
 };
 
-@runsWhen([pageDetect.isRepoHome])
-class RepoActiInflTrend extends PerceptorBase {
-  private _currentRepo: string;
-
-  constructor() {
-    super();
-    this._currentRepo = '';
-  }
-
-  public async run(): Promise<void> {
-    this._currentRepo = utils.getRepositoryInfo(window.location)!.nameWithOwner;
-
-    let newBorderGridRow = null;
-    let newBorderGridCell = null;
-
-    // if not the first time to enter this code
-    if (document.getElementById('repo-acti-infl-trend') != null) {
-      newBorderGridCell = $('#repo-acti-infl-trend').children(
-        '.BorderGrid-cell'
-      )[0];
-
-      render(
-        <RepoActiInflTrendView currentRepo={this._currentRepo} />,
-        newBorderGridCell
-      );
-    } else {
-      newBorderGridRow = document.createElement('div');
-      newBorderGridRow.id = 'repo-acti-infl-trend';
-      newBorderGridRow.className = 'BorderGrid-row';
-      newBorderGridCell = document.createElement('div');
-      newBorderGridCell.className = 'BorderGrid-cell';
-      newBorderGridRow.appendChild(newBorderGridCell);
-
-      render(
-        <RepoActiInflTrendView currentRepo={this._currentRepo} />,
-        newBorderGridCell
-      );
-
-      const borderGridRows = $('div.Layout-sidebar').children('.BorderGrid');
-      borderGridRows.append(newBorderGridRow);
-    }
-  }
-}
-
-inject2Perceptor(RepoActiInflTrend);
+export default RepoActiInflTrendView;
